lab1b fixes
