--- README.md
@@ -1,176 +1,0 @@
-# [WIP] Лабораторные работы за осенний семестр 2020
-
-В `master` ветке данного репозитория присутствуют актуальные задания, включая 
-распределение протоколов по студентам.
-
-# Прогресс
-
-Легенда:
-
-| Символ | Значение                                      |
-| --     | --                                            |
-|        | Работа не сдана или находится на рассмотрении |
-| +      | Работа принята                                |
-| !      | Работа рассмотрена, требуется демо            |
-| ?      | Работа рассмотрена, но требует доработки      |
-| -      | Работу сдавать не требуется                   |
-
-
-## Группа 201
-
-| ФИО              | Ssh ключ | Лаб.1 | Лаб.1б | Лаб.2 | Лаб.3 |
-| --               | --       | --    | --     | --    | --    |
-| Антропова А.А.   | +        |       |        |       |       |
-| Белов Е.А.       | +        |       |        |       |       |
-| Буй К.Д.         | +        |       |        |       |       |
-| Гладкова Е.Д.    | +        |       |        |       |       |
-| Голзицкий Н.С.   | +        |       |        |       |       |
-| Гуляев Д.В.      |          |       |        |       |       |
-| Данилов А.И.     | +        |       |        |       |       |
-| Казанджи М.А.    | +        |       |        |       |       |
-| Киселев Н.Д.     | +        |       |        |       |       |
-| Лялин А.С.       | +        |       |        |       |       |
-| Натура А.А.      |          |       |        |       |       |
-| Никитин И.Н.     | +        |       |        |       |       |
-| Романов А.Л.     | +        |       |        |       |       |
-| Свечников Р.А.   | +        |       |        |       |       |
-| Сибагатулин А.Ф. |          |       |        |       |       |
-| Товпеко К.А.     | +        |       |        |       |       |
-| Черноног С.А.    | +        |       |        |       |       |
-| Шаляпин Г.А.     | +        |       |        |       |       |
-
-## Группа 203
-
-| ФИО             | Ssh ключ | Лаб.1 | Лаб.1б | Лаб.2 | Лаб.3 |
-| --              | --       | --    | --     | --    | --    |
-| Ворошилов А.А.  | +        |       |        |       |       |
-| Гусев Н.С.      |          |       |        |       |       |
-| Зарецкая Е.С.   |          |       |        |       |       |
-| Иванов И.Д.     | +        |       |        |       |       |
-| Калашников Р.А. |          |       |        |       |       |
-| Костарев В.И.   | +        |       |        |       |       |
-| Любченкова А.А. | +        |       |        |       |       |
-| Меньшов П.А.    | +        |       |        |       |       |
-| Морозов Е.С.    | +        |       |        |       |       |
-| Никитина Д.С.   | +        |       |        |       |       |
-| Овсянников Е.А. | +        |       |        |       |       |
-| Орлова П.А.     | +        |       |        |       |       |
-| Пентегов А.О.   |          |       |        |       |       |
-| Семенов Д.С.    | +        |       |        |       |       |
-| Середин К.В.    | +        |       |        |       |       |
-| Трушин И.А.     | +        |       |        |       |       |
-| Черникова А.С.  | +        |       |        |       |       |
-| Шелепов В.А.    | +        |       |        |       |       |
-
-# Требования к отчету:
-
-* Инструкция по использованию;
-* Инструкция по сборке/установке;
-* Описание используемого протокола;
-  * Своего для первой лабораторной;
-  * Используемого подмножества для второй;
-* ???
-* PROFIT
-
-Отчет можно писать как в сообщении к PR-у, так и присылать в иных 
-форматах (`.pdf`, `.docx`, `.txt` и т.п.).
-
-# Порядок сдачи
-
-* Fork от данного репозитория
-* Push каждой лабораторной в отдельную ветку
-* Создание отдельного PR на каждую лабораторную
-* Ставите label по сдаваемой лабе (`First lab`, `Second lab`, `Third lab`)
-
-При создании PR, в качестве напоминалки, сделан шаблон.
-
-
-## Лабораторная 1.а
-
-Данная работа нацелена на реализацию чата.
-Разрабатывается серверная и клиентская части.
-
-Требования к чату:
-
-* Возможность подключения нескольких клиентов к одному чату;
-* Отображение отправленного сообщения у каждого подключенного в этот
-момент к серверу клиента.
-
-Требования к реализации:
-
-* Самостоятельное использование TCP сокетов
-
-Реализация хранения и отображения истории **не обязательно**.
-Используемый язык программирования - **на выбор студента** (никаких ограничений)
-
-Формат сообщения следующий:
-
-```
-<HH:MM> [username] Text of an actual message
-```
-
-Пример сообщения:
-
-```
-<4:20> [Kenobi] Hello there!
-```
-
-## Лабораторная 1.б
-
-Всё как и в первой, но на неблокирующих сокетах.
-
-## Лабораторная 2
-
-Вторая лабораторная нацелена на реализацию протоколов, работающих по UDP.
-Разрабатывается серверная и клиентская части для разных протоколов (серверная -
-один протокол, клиентская - другой).
-
-Язык программирования - **на выбор студента**
-
-Реализованный студентом протокол будет тестироваться с имеющимися реализациями
-(при реализации сервера - тестируем с уже написанным кем-то клиентом, при 
-реализации клиента - наоборот).
-
-Для изменения/назначения своих сервера и клиента можно написать в чат своей группы по данному предмету или завести отдельный PR с изменением нижеприведённой таблицы.
-
-### Группа 201
-
-| Студент              | Сервер      | Клиент      |
-| -------------------- | ----------- | ----------- |
-| Антропова Анна       |             |             |
-| Белов Егор           |             |             |
-| Буй Куок Дат         |             |             |
-| Гладкова Екатерина   |             |             |
-| Голзицкий Николай    |             |             |
-| Гуляев Дмитрий       |             |             |
-| Данилов Алексей      | dhcp        | ntp         |
-| Казанджи Михаил      | ntp         | snmp        |
-| Киселев Никита       | dns         | tftp        |
-| Лялин Александр      |             |             |
-| Натура Антон         |             |             |
-| Никитин Илья         |             |             |
-| Романов Алексей      | dns(dhcp)   | dhcp(dns)   |
-| Свечников Роман      | dhcp(dns)   | dns(dhcp)   |
-| Сибагатулин Азат     |             |             |
-| Товпеко Кирилл       | dns         | ntp         |
-| Черноног Сергей      |             |             |
-| Шаляпин Георгий      |             |             |
-
-
-### Группа 203
-
-| Студент     | Сервер | Клиент |
-| --          | --     | --     |
-| Трушин Иван | tftp   | dhcp   |
-| Шелепов     | dhcp   | tftp   |
-| Меньшов     | dns    | tftp   |
-| Морозов     | dns    | dhcp   |
-| Орлова      | tftp   | dns    |
-| Семёнов     | dhcp   | dns    |
-| Калашников  | dns    | snmp   |
-| Костарев    | dhcp   | snmp   |
-| Середин     | dhcp   | ntp    |
-
-
-## Лабораторная 3
-
