--- messengerb/command.txt
@@ -1,2 +1,0 @@
-gcc client.c -o client `pkg-config --cflags gtk+-3.0` `pkg-config --libs gtk+-3.0`
-
